# dist/swagger.yaml
openapi: 3.0.1
info:
  title: Backend API For the TaskManagementSystem
  description: Backend API For the TaskManagementSystem
  license:
    name: Apache 2.0
    url: 'http://springdoc.org'
  version: v0.0.1
...
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
paths:
...
  '/api/comment/add/{id}':
    post:
      tags:
        - comment-controller
      summary: Add Comment to Task
      operationId: addCommentToTask
      parameters:
        - name: id
          in: path
          description: Task id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInputDto'
      responses:
        '200':
          description: Add Comment to Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          description: 'If RequestBody, {id} is missing'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: If Task not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
...
  '/api/task/{id}':
    get:
      tags:
        - task-controller
      summary: Get Task
      operationId: getTask
      parameters:
        - name: id
          in: path
          description: Task id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Get Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          description: 'If {id} is missing'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: If Task not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
...
  '/api/comment/{id}':
    get:
      tags:
        - comment-controller
      summary: Get Comment
      operationId: getComment
      parameters:
        - name: id
          in: path
          description: Comment id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Get Comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          description: 'If {id} is missing'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: If Comment not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
  '/api/comment/task/{id}':
    get:
      tags:
        - comment-controller
      summary: Get All Task's Comments
      operationId: getCommentsByTask
      parameters:
        - name: id
          in: path
          description: Task id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
      responses:
        '200':
          description: Get All Task's Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentDto'
        '400':
          description: 'If {id} is missing, page < 0'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: If page not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
...
components:
  schemas:
    ApiErrorDto:
      required:
        - created
        - errors
        - instance
        - status
        - title
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
        instance:
          type: string
        created:
          type: string
          format: date-time
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDto'
    ErrorDto:
      required:
        - description
        - error
      type: object
      properties:
        error:
          type: string
        description:
          type: string
...
    StatusDto:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 32
          minLength: 1
          type: string
    PriorityDto:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 32
          minLength: 1
          type: string
    TaskDto:
      required:
        - author
        - createdOn
        - description
        - executor
        - id
        - lastUpdatedOn
        - priority
        - status
        - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        author:
          $ref: '#/components/schemas/UserDto'
        executor:
          $ref: '#/components/schemas/UserDto'
        status:
          $ref: '#/components/schemas/StatusDto'
        priority:
          $ref: '#/components/schemas/PriorityDto'
        createdOn:
          type: string
          format: date-time
        lastUpdatedOn:
          type: string
          format: date-time
    UserDto:
      required:
        - createdOn
        - email
        - username
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        createdOn:
          type: string
          format: date-time
    CommentInputDto:
      required:
        - text
      type: object
      properties:
        text:
          maxLength: 1024
          minLength: 10
          type: string
    CommentDto:
      required:
        - author
        - createdOn
        - id
        - lastUpdatedOn
        - text
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
        author:
          $ref: '#/components/schemas/UserDto'
        createdOn:
          type: string
          format: date-time
        lastUpdatedOn:
          type: string
          format: date-time
...